rc2

+++ lvmluks.sh
@@ -13,20 +13,24 @@
 
 # Written by cCore@freenode
 
-echo -e 'Select the disk you want to prepare for luks and lvm'
-ls -l /dev/[sh]d[a-z]
-read -e disk
 
+if [ "$(whoami)" != 'root' ]; then
+        echo 'You need to be root.'
+        exit 1;
+fi
+echo -e 'Select the disk you want to prepare for luks and lvm'
+	ls -l /dev/[sh]d[a-z]
+	read -e disk
 echo -e 'Would you like to fill the disk with random data?\nNOTE: This will take a while! But offers better cryptoanalytical integrity.\nY/N?'
-read = fill
-	if [ fill = 'y' ]; then
+	read = fill
+if [ fill = 'y' ]; then
 	dd if=/dev/urandom of=$disk
 fi
-
 echo -e '\nNow we will use fdisk to partition the drive and create 2 partitions.'
 echo -e '1st partition will be used as /boot and will not be encrypted.'
 echo -e '2nd partition will be the remainder of the disk, and we will use this for the lvm encrypted part.\n'
-read -n1 -r -p "Press any key to continue..."
+	read -n1 -r -p "Press any key to continue..."
+
 fdisk $disk <<EOF
 d
 4
@@ -65,7 +69,6 @@ cryptsetup -s 256 -y luksFormat $diskpart
 sleep 2
 cryptsetup luksOpen $diskpart crypt
 pvcreate /dev/mapper/crypt
-
 echo -e '\nWould you like to set your own name on the volume group?\nIf not, i will create one called vgcrypt\nY/N?'
 	read -e askvg
 if [ $askvg = 'y' ]; then
@@ -82,7 +85,6 @@ echo -e Chose your home partition size. eg. 10G/1024M
 echo -e Chose your swap partition size. eg. 10G/1024M
 	read -r swap
 		lvcreate -L $swap -n swap vgenc
-
 echo -e '\nWould you like to create additional partitions?\nY/N'
 read -e ask
 while [ $ask = 'y' ]; do
@@ -90,14 +92,12 @@ while [ $ask = 'y' ]; do
 		read -e newpart
 	echo -e 'Enter your '$newpart' partition size. eg. 10G/1024M'
 		read -e newsize
-		lvcreate -L $newsize -n $newsize vgenc
+		lvcreate -L $newsize -n $newpart vgenc
 	echo -e 'Add more?\nY/N'
 		read -e ask
 done
-
 vgscan --mknodes
 vgchange -ay
 mkswap /dev/vgenc/swap
 cryptsetup luksDump $disk
-
 echo -e '\nAll done.\nYou can now run setup.'
